+++ src/runtime/src/core/num4.rs
@@ -1,9 +1,14 @@
-use crate::{runtime_error, Value, ValueMap};
+use crate::{runtime_error, Value, ValueIterator, ValueMap};
 
 pub fn make_module() -> ValueMap {
     use Value::*;
 
     let mut result = ValueMap::new();
+
+    result.add_fn("iter", |vm, args| match vm.get_args(args) {
+        [Num4(n)] => Ok(Iterator(ValueIterator::with_num4(*n))),
+        _ => runtime_error!("num4.iter: Expected a Num4 as argument"),
+    });
 
     result.add_fn("sum", |vm, args| match vm.get_args(args) {
         [Num4(n)] => Ok(Number(